Blueprint v3 - mostly done and working but without the collision handling.
The new insight is that short hashes will not work for pushdata_hashes so need to rework it
to index full pushdata_hashes in a separate table (with uniqueness constraint) and an auto-incrementing
pushdata_num for use in the main pushdata table (where the 8 byte pushdata_num will be repeated many times
over rather than the full 32 byte hash).
also added pyproject.toml and increased "black" code formatting to line-length 100.