Checkpoint. Proceed with refactoring the pipeline. (status of main issues below)
Shutdown is working mostly (but with a few noisy exceptions from child processes) - it still needs work.
TODO - move termination of logging server back out to top level main startup script to capture the final shutdown logging output
Need to debug why it froze at block 139443 with some blocks that have incomplete counts (controller still waiting on processing but not getting them)
I am not sure exactly the cause of the pipeline sometimes blocking up. But with the refactoring it should become more clear and easier to diagnose.

Plan now is to join:
- Controller + Preprocessor + Raw Blocks + MerkleTree together into "ConduitRaw"
- Controller + TxParsers into "ConduitIdx" which will re-use most of the same controller logic
- There should be a general move away from shared memory and towards ZMQ binary message passing instead for all IPC.